config: clean up cloudinit functions
