+++ src/config/config.go
@@ -16,11 +16,9 @@ package config
 
 import (
 	"bytes"
-	"compress/gzip"
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io/ioutil"
 
 	"github.com/coreos/ignition/config/types"
 
@@ -41,9 +39,9 @@ func Parse(rawConfig []byte) (config types.Config, err error) {
 		}
 	} else if isEmpty(rawConfig) {
 		err = ErrEmpty
-	} else if isCloudConfig(decompressIfGzipped(rawConfig)) {
+	} else if isCloudConfig(rawConfig) {
 		err = ErrCloudConfig
-	} else if isScript(decompressIfGzipped(rawConfig)) {
+	} else if isScript(rawConfig) {
 		err = ErrScript
 	}
 	if serr, ok := err.(*json.SyntaxError); ok {
@@ -56,18 +54,4 @@ func Parse(rawConfig []byte) (config types.Config, err error) {
 
 func isEmpty(userdata []byte) bool {
 	return len(userdata) == 0
-}
-
-func decompressIfGzipped(data []byte) []byte {
-	if reader, err := gzip.NewReader(bytes.NewReader(data)); err == nil {
-		uncompressedData, err := ioutil.ReadAll(reader)
-		reader.Close()
-		if err == nil {
-			return uncompressedData
-		} else {
-			return data
-		}
-	} else {
-		return data
-	}
 }
